requirements.txt
##Libraries to installl

xlwt
opencv-python
numpy
matplotlib
pillow
pathlib
MIME
